Tweak r24911 for https://trac.osgeo.org/gdal/ticket/4803 -- don't escape DESDATA for sizes >10mb
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@24914 f0d54148-0727-0410-94bb-9a71ac55c965

+++ frmts/nitf/nitfdes.c
@@ -280,7 +280,7 @@ retry:
 
 #define TEN_MEGABYTES 10485760
             
-        if ((int)psSegInfo->nSegmentSize > TEN_MEGABYTES)
+        if (psSegInfo->nSegmentSize > TEN_MEGABYTES)
         {
             char* pszOffset = CPLSPrintf(CPL_FRMT_GUIB, psFile->pasSegmentInfo[iSegment].nSegmentStart);
             char* pszSize = CPLSPrintf(CPL_FRMT_GUIB, psFile->pasSegmentInfo[iSegment].nSegmentSize);
